+++ app/Http/Controllers/Api/v1/CategoryController.php
@@ -9,6 +9,8 @@ use App\Http\Requests\Category\UpdateCategoryRequest;
 use App\Http\Repositories\v1\CategoryRepository;
 use Illuminate\Http\JsonResponse;
 use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
+
 /**
  * @group Category
  *
@@ -16,18 +18,16 @@ use App\Models\Category;
 class CategoryController extends Controller
 {
 
-    public function __construct(public CategoryRepository $categoryRepository)
-    {
-    }
+    public function __construct(public CategoryRepository $categoryRepository) {}
 
     /**
-    * Category Get all
-    *
-    * @response {
+     * Category Get all
+     *
+     * @response {
     {{response}}
-    * }
-    * @return JsonResponse
-    */
+     * }
+     * @return JsonResponse
+     */
 
     public function index(Request $request)
     {
@@ -35,31 +35,32 @@ class CategoryController extends Controller
     }
 
     /**
-    * Category adminIndex get All
-    *
-    * @response {
+     * Category adminIndex get All
+     *
+     * @response {
     {{response}}
-    * }
-    * @return JsonResponse
-    */
+     * }
+     * @return JsonResponse
+     */
 
     public function adminIndex(Request $request)
     {
+        Gate::authorize('categories.view');
         return $this->categoryRepository->adminIndex($request);
     }
 
     /**
-    * Category view
-    *
-    * @queryParam id required
-    *
-    * @param Request $request
-    * @param int     $id
-    * @return JsonResponse
-    * @response {
+     * Category view
+     *
+     * @queryParam id required
+     *
+     * @param Request $request
+     * @param int     $id
+     * @return JsonResponse
+     * @response {
     {{response}}
-    * }
-    */
+     * }
+     */
 
     public function show(Request $request, Category $category): JsonResponse
     {
@@ -67,38 +68,40 @@ class CategoryController extends Controller
     }
 
     /**
-    * Category create
-    *
-         * @bodyParam name string
+     * Category create
+     *
+     * @bodyParam name string
      * @bodyParam parent_id integer
 
-    *
-    * @param StoreCategoryRequest $request
-    * @return JsonResponse
-    */
+     *
+     * @param StoreCategoryRequest $request
+     * @return JsonResponse
+     */
 
     public function store(StoreCategoryRequest $request): JsonResponse
     {
+        Gate::authorize('categories.create');
         return $this->categoryRepository->store($request);
     }
 
     /**
-    * Category update
-    *
-    * @queryParam category required
-    *
-         * @bodyParam name string
+     * Category update
+     *
+     * @queryParam category required
+     *
+     * @bodyParam name string
      * @bodyParam parent_id integer
 
-    *
-    * @param UpdateCategoryRequest $request
-    * @param Category $category
-    * @return JsonResponse
-    */
+     *
+     * @param UpdateCategoryRequest $request
+     * @param Category $category
+     * @return JsonResponse
+     */
 
     public function update(UpdateCategoryRequest $request, Category $category): JsonResponse
     {
-         return $this->categoryRepository->update($request, $category);
+        Gate::authorize('categories.update');
+        return $this->categoryRepository->update($request, $category);
     }
 
     /**
@@ -112,6 +115,7 @@ class CategoryController extends Controller
 
     public function destroy(Category $category): JsonResponse
     {
-        return  $this->categoryRepository->destroy($category);
+        Gate::authorize('categories.delete');
+        return $this->categoryRepository->destroy($category);
     }
 }
