+++ app/Http/Controllers/Api/v1/QuestionController.php
@@ -9,6 +9,8 @@ use App\Http\Requests\Question\UpdateQuestionRequest;
 use App\Http\Repositories\v1\QuestionRepository;
 use Illuminate\Http\JsonResponse;
 use App\Models\Question;
+use Illuminate\Support\Facades\Gate;
+
 /**
  * @group Question
  *
@@ -16,18 +18,16 @@ use App\Models\Question;
 class QuestionController extends Controller
 {
 
-    public function __construct(public QuestionRepository $questionRepository)
-    {
-    }
+    public function __construct(public QuestionRepository $questionRepository) {}
 
     /**
-    * Question Get all
-    *
-    * @response {
+     * Question Get all
+     *
+     * @response {
     {{response}}
-    * }
-    * @return JsonResponse
-    */
+     * }
+     * @return JsonResponse
+     */
 
     public function index(Request $request)
     {
@@ -35,31 +35,33 @@ class QuestionController extends Controller
     }
 
     /**
-    * Question adminIndex get All
-    *
-    * @response {
+     * Question adminIndex get All
+     *
+     * @response {
     {{response}}
-    * }
-    * @return JsonResponse
-    */
+     * }
+     * @return JsonResponse
+     */
 
     public function adminIndex(Request $request)
     {
+        Gate::authorize('questions.view');
+
         return $this->questionRepository->adminIndex($request);
     }
 
     /**
-    * Question view
-    *
-    * @queryParam id required
-    *
-    * @param Request $request
-    * @param int     $id
-    * @return JsonResponse
-    * @response {
+     * Question view
+     *
+     * @queryParam id required
+     *
+     * @param Request $request
+     * @param int     $id
+     * @return JsonResponse
+     * @response {
     {{response}}
-    * }
-    */
+     * }
+     */
 
     public function show(Request $request, Question $question): JsonResponse
     {
@@ -67,42 +69,46 @@ class QuestionController extends Controller
     }
 
     /**
-    * Question create
-    *
-         * @bodyParam level_id integer
+     * Question create
+     *
+     * @bodyParam level_id integer
      * @bodyParam category_id integer
      * @bodyParam question_text string
      * @bodyParam type integer
 
-    *
-    * @param StoreQuestionRequest $request
-    * @return JsonResponse
-    */
+     *
+     * @param StoreQuestionRequest $request
+     * @return JsonResponse
+     */
 
     public function store(StoreQuestionRequest $request): JsonResponse
     {
+        Gate::authorize('questions.create');
+
         return $this->questionRepository->store($request);
     }
 
     /**
-    * Question update
-    *
-    * @queryParam question required
-    *
-         * @bodyParam level_id integer
+     * Question update
+     *
+     * @queryParam question required
+     *
+     * @bodyParam level_id integer
      * @bodyParam category_id integer
      * @bodyParam question_text string
      * @bodyParam type integer
 
-    *
-    * @param UpdateQuestionRequest $request
-    * @param Question $question
-    * @return JsonResponse
-    */
+     *
+     * @param UpdateQuestionRequest $request
+     * @param Question $question
+     * @return JsonResponse
+     */
 
     public function update(UpdateQuestionRequest $request, Question $question): JsonResponse
     {
-         return $this->questionRepository->update($request, $question);
+        Gate::authorize('questions.update');
+
+        return $this->questionRepository->update($request, $question);
     }
 
     /**
@@ -116,6 +122,8 @@ class QuestionController extends Controller
 
     public function destroy(Question $question): JsonResponse
     {
+        Gate::authorize('questions.delete');
+
         return  $this->questionRepository->destroy($question);
     }
 }
